requirements-step1.txt
langgraph>=0.2.40
neo4j>=5.23.0
bertopic>=0.16.3
sentence-transformers>=3.0.1
transformers>=4.44.0
torch>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
APScheduler>=3.10.4
feedparser>=6.0.11
python-dateutil>=2.9.0
pydantic>=2.8.2
